+++ policyfiles/cron.rb
@@ -6,4 +6,4 @@ run_list 'managed-chef-server::cron'
 
 # every 5 minutes for testing
 override['mcs']['cron']['minute'] = '*/5'
-override['mcs']['cron']['options'] = ['-z']
+override['mcs']['cron']['options'] = ['-z', '-F min']