+++ CHANGELOG.md
@@ -28,3 +28,4 @@ This file is used to list changes made in each version of the managed-chef-serve
 
 - refactor policyfiles for more straightforward testing
 - backup scheduled via cron and attributes
+- cron recipe for managing the chef-server with the chef-client under cron, with or without a policyfile archive